Use modified observer pattern
Entries Register::register() whenever they are changed
Manifest Register::subscribe() on __construct()
Register::register() calls Manifest updateEntryStatus($uuid, array$entry) with changed entry

doing it this way so that 'dangling' entry instances can't lead to lagging changes
e.g.    $entry = $manifest->getEntry($uuid)//Entry instance
        $entry->makeSomeWhateverChanges()//make some changes to entry
        $manifest->getEntry($uuid)->getSomeWhateverChanges()//oops, no changes! different entry
that's how it was, because Entry instances only Register::register()ed on __destruct()
so if an instance was left hanging around somewhere and you tried to access it again through Manifest
which is essentially what Manifest is for, any changes to it wouldn't show up.
This is to correct that.

The reason for not just using the same Entry objects all the time is that there are potentially
a lot of entries to be passed around but creating lots of objects makes responses slow.
associative arrays work just fine for querying or representing entries in a JSON response
and they're fast. But we still want the convenience of Entry methods like getImage, moveFile,
or setAttribute where Entries manage attribute names for us. So Manifest is responsible for
passing around Entry instances. Entries are responsible for all the data and meta data concerning a photo;
where it's kept in the file system, what it's url is, what sort of image it is, tracking its hash identifier
etc etc, and it can expand functionality without Manifest needing to know much or anything about it. Register
is responsible for making sure that Entries persist in the filesystem, that attributes and meta data
associated with them are reported to the Manifest, saved after modification, and restored upon the next request.
So Manifest is evolving into an interface for the model that is maintained by Entries and Register.