remove Entry methods getAttributeKeys and getPropertyKeys
since they're agregation methods and you only ever want all the keys from
all the entries these methods had to be run through another method in Manifest
anyway to filter out duplicates, but it's much faster for manifest to operate
on an array than to create a bunch of objects just to call these methods.

add Manifest::getFile($uuid) returns the filepath of a file with basename of $uuid
so when just a photo is requested, not a collection of photo resources, Manifest
can bypass the whole json decode to array instantiate Entry process and just return
the image file directly so the photo server can send it out pronto

modify Register __destruct to check that Register::entries is actually an array
before json encoding it and writing it to the manifest.json file, so if something goes
wrong it doesn't just overwrite all the data with null.

modify Register __construct to get rid of the throw statement when json isn't decoded properly
there isn't anything to be done about it on any level, now it just makes sure it's
passing an array so it doesn't cause an error cascade by passing the wrong type
and the application just assumes there are no entries.